+++ components/browser/state/src/test/java/mozilla/components/browser/state/engine/EngineObserverTest.kt
@@ -86,6 +86,16 @@ class EngineObserverTest {
                 onResult: (JSONObject) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
+            override fun sendGleanBrokenSiteReport(
+                details: JSONObject?,
+                description: String?,
+                reason: String,
+                url: String,
+                sendTabSpecificInfo: Boolean,
+                sendBlockedUrls: Boolean,
+                onResult: () -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
             override fun getWebCompatInfo(
                 onResult: (JSONObject) -> Unit,
                 onException: (Throwable) -> Unit,
@@ -180,6 +190,16 @@ class EngineObserverTest {
             ) {}
             override fun getBrokenSiteReport(
                 onResult: (JSONObject) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
+            override fun sendGleanBrokenSiteReport(
+                details: JSONObject?,
+                description: String?,
+                reason: String,
+                url: String,
+                sendTabSpecificInfo: Boolean,
+                sendBlockedUrls: Boolean,
+                onResult: () -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
             override fun getWebCompatInfo(
@@ -277,6 +297,16 @@ class EngineObserverTest {
                 onResult: (JSONObject) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
+            override fun sendGleanBrokenSiteReport(
+                details: JSONObject?,
+                description: String?,
+                reason: String,
+                url: String,
+                sendTabSpecificInfo: Boolean,
+                sendBlockedUrls: Boolean,
+                onResult: () -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
             override fun getWebCompatInfo(
                 onResult: (JSONObject) -> Unit,
                 onException: (Throwable) -> Unit,
@@ -368,6 +398,16 @@ class EngineObserverTest {
                 onResult: (JSONObject) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
+            override fun sendGleanBrokenSiteReport(
+                details: JSONObject?,
+                description: String?,
+                reason: String,
+                url: String,
+                sendTabSpecificInfo: Boolean,
+                sendBlockedUrls: Boolean,
+                onResult: () -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
             override fun getWebCompatInfo(
                 onResult: (JSONObject) -> Unit,
                 onException: (Throwable) -> Unit,
@@ -455,6 +495,16 @@ class EngineObserverTest {
             ) {}
             override fun getBrokenSiteReport(
                 onResult: (JSONObject) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
+            override fun sendGleanBrokenSiteReport(
+                details: JSONObject?,
+                description: String?,
+                reason: String,
+                url: String,
+                sendTabSpecificInfo: Boolean,
+                sendBlockedUrls: Boolean,
+                onResult: () -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
             override fun getWebCompatInfo(