+++ components/support/test-fakes/src/main/java/mozilla/components/support/test/fakes/engine/TestEngineSession.kt
@@ -67,6 +67,17 @@ open class TestEngineSession(override val settings: Settings = DefaultSettings()
         onException: (Throwable) -> Unit,
     ) = Unit
 
+    override fun sendGleanBrokenSiteReport(
+        details: JSONObject?,
+        description: String?,
+        reason: String,
+        url: String,
+        sendTabSpecificInfo: Boolean,
+        sendBlockedUrls: Boolean,
+        onResult: () -> Unit,
+        onException: (Throwable) -> Unit,
+    ) = Unit
+
     override fun getWebCompatInfo(
         onResult: (JSONObject) -> Unit,
         onException: (Throwable) -> Unit,